Iframe: simplify/reactify compat styles logic
